fix: fixed adding comment
