+++ frontend/appflowy_web_app/src/components/global-comment/add-comment/AddCommentWrapper.tsx
@@ -1,0 +1,60 @@
+import { getScrollParent } from '@/components/global-comment/utils';
+import { HEADER_HEIGHT } from '@/components/publish/header';
+import React, { useEffect, useRef, useState } from 'react';
+import AddComment from './AddComment';
+import { Portal } from '@mui/material';
+
+export function AddCommentWrapper() {
+  const addCommentRef = useRef<HTMLDivElement>(null);
+  const [showFixedAddComment, setShowFixedAddComment] = useState(false);
+  const [focus, setFocus] = useState(false);
+  const [content, setContent] = useState('');
+
+  useEffect(() => {
+    const element = addCommentRef.current;
+
+    if (!element) return;
+    const scrollContainer = getScrollParent(element);
+
+    if (!scrollContainer) return;
+
+    const handleScroll = () => {
+      const isIntersecting = element.getBoundingClientRect().top < HEADER_HEIGHT;
+
+      if (isIntersecting) {
+        setShowFixedAddComment(true);
+      } else {
+        setShowFixedAddComment(false);
+      }
+    };
+
+    scrollContainer.addEventListener('scroll', handleScroll);
+    return () => {
+      scrollContainer.removeEventListener('scroll', handleScroll);
+    };
+  }, []);
+
+  return (
+    <>
+      <div className={'my-2'} ref={addCommentRef}>
+        <AddComment
+          content={content}
+          setContent={setContent}
+          focus={focus && !showFixedAddComment}
+          setFocus={setFocus}
+        />
+      </div>
+      {showFixedAddComment && (
+        <Portal container={document.body}>
+          <div className={'fixed top-[48px] flex w-full justify-center'}>
+            <div className={'w-[964px] min-w-0 max-w-full px-16 max-sm:px-4'}>
+              <AddComment content={content} setContent={setContent} focus={focus} setFocus={setFocus} />
+            </div>
+          </div>
+        </Portal>
+      )}
+    </>
+  );
+}
+
+export default AddCommentWrapper;
