+++ frontend/appflowy_web_app/src/components/global-comment/add-comment/index.ts
@@ -1,0 +1,1 @@
+export * from './AddCommentWrapper';
