+++ frontend/appflowy_web_app/src/components/global-comment/reactions/Reactions.tsx
@@ -1,4 +1,5 @@
 import { GlobalComment, Reaction as ReactionType } from '@/application/comment.type';
+import ReactAction from '@/components/global-comment/actions/ReactAction';
 import { useGlobalCommentContext } from '@/components/global-comment/GlobalComment.hooks';
 import Reaction from '@/components/global-comment/reactions/Reaction';
 import React, { memo, useCallback, useMemo } from 'react';
@@ -22,6 +23,7 @@ export function Reactions({ comment }: { comment: GlobalComment }) {
       {commentReactions.map((reaction) => {
         return <Reaction reaction={reaction} onClick={handleReactionClick} key={reaction.reactionType} />;
       })}
+      <ReactAction comment={comment} />
     </div>
   );
 }
