+++ frontend/appflowy_web_app/src/components/global-comment/utils.ts
@@ -1,0 +1,11 @@
+export function getScrollParent(node: Element | null): Element | null {
+  if (node === null) {
+    return null;
+  }
+
+  if (node.scrollHeight > node.clientHeight) {
+    return node;
+  } else {
+    return getScrollParent(node.parentElement);
+  }
+}
