+++ frontend/appflowy_web_app/src/styles/app.scss
@@ -150,3 +150,7 @@ body {
   box-shadow: 0px -15px 20px 12px var(--bg-body);
   background-color: transparent;
 }
+
+.comment {
+  scroll-margin-top: 100px;
+}
